+++ NEWS.md
@@ -76,3 +76,6 @@
 * Add methods to calculate point leverage, the effect of removing different strains and sera from the map, `agLeverage()`, `srLeverage()`, `titerLeverage()`
 * Change the `view.acmap()` method to allow for an option `num_optimizations` to specify how many optimizations to send to the viewer.
 * Add an option for and basic check by default for convergence of solutions when running map optimizations.
+
+# Racmacs 1.1.20
+* Slight re-implementation of dimensional annealing to not report progress bar multiple times.